Revert "Tulispelliin liittyviä koodeja"
