--- Hunt for Magic/Assets/Artcorner.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 59b5c6253ece52b47b9ddca3c0774783
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
