--- Hunt for Magic/Assets/Artcorner/New Material.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f26cd90fef72b9348af204f531cb11b4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
