--- Hunt for Magic/Assets/Prefabs/Flamethrower_particle.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7b2c73991c1b38145a3e6d5387f918c8
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
