+++ Hunt for Magic/Assets/Prefabs/WindSpellPrefab.prefab
@@ -30,7 +30,7 @@ Transform:
   m_GameObject: {fileID: 5189080076824686935}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -1.801, y: 0.432, z: 1.737}
-  m_LocalScale: {x: 0.2, y: 0.2, z: 0.2}
+  m_LocalScale: {x: 0.1, y: 0.1, z: 0.1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
@@ -106,7 +106,7 @@ Rigidbody:
   m_Mass: 1
   m_Drag: 0
   m_AngularDrag: 0
-  m_UseGravity: 0
+  m_UseGravity: 1
   m_IsKinematic: 0
   m_Interpolate: 0
   m_Constraints: 0
@@ -123,4 +123,3 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 4d4448a297e39f246a7503f66bc35b1a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _damageAmount: 10
