--- Hunt for Magic/Assets/Scenes/Artgallery.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 34cb4d9e07bebe344be3af4c92700dab
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
