--- Hunt for Magic/Assets/Scripts/EnergySystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bb683163f853f2d47bf8eed1bb52f1d1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
