--- Hunt for Magic/Assets/Scripts/FireSpell.cs
@@ -1,34 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class FireSpell : MonoBehaviour
-{
-
-    [SerializeField]
-    private float _damageAmount = 0.01f;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        Destroy(gameObject, 7f);
-    }
-
-    private void OnParticleCollision(GameObject other)
-    {
-        var enemy = other.gameObject.GetComponent<Rigidbody>();
-
-        var enemyHealth = other.gameObject.GetComponent<HealthSystem>();
-
-        if (enemy != null && enemy.name != "PlayerCharacter")
-        {
-            enemyHealth.AddDamage(_damageAmount);
-        }
-    }
-}
