--- Hunt for Magic/Assets/Scripts/FireSpell.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9290204589e0aeb42a4a76032e64f157
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
